Copepod fish interaction database

A list of interactions between copepods and fish based on WoRMS database. And curated by Francisco Neptalí Morales-Serna. This dataset was cleaned and formated by Angel Luis Robles Fernandez following the GloBi (Global interactions) database standard. The cleaned dataset is stored in 10.5281/zenodo.11002619

Usage

cofid

Format

## `cofid` A data frame with 10, 334 rows and 9 columns:

source_taxon_external_id

WoRMS id of the parasite, i.e. AphiaID

source_taxon_name

Scientific name of the species

source_taxon_path

Taxon path of the parasite

source_specimen_life_stage

Life stage of the parasite

interaction_type

Type of interactio between parasite and host

target_taxon_external_id

WoRMS id of the host, i.e. AphiaID

target_taxon_name

Scientific name of the host

target_taxon_path

Taxon path of the host

study_citation

Source paper of the interaction

Source

Copepod fish interaction database <10.5281/zenodo.11511665>