Find copepod occurrence data for given copepod species. Given a copepod family, genus or species returns a list of copepod occurrences.

Find copepod occurrence data for given copepod species. Given a copepod family, genus or species returns a list of copepod occurrences.

Usage

getCopepods(species = NULL, genus = NULL, family = NULL, citation = TRUE)

Arguments

species

Copepod species to filter

genus

Copepod genus to filter

family

Copepod family to filter

citation

boolean. Should the output include the citation information? default is FALSE

Value

A data.frame with copepod fish interactions

Examples

getCopepods(genus = "Caligus")
#> # A tibble: 1,864 × 3
#>    source_taxon_name target_taxon_name   study_citation                         
#>    <chr>             <chr>               <chr>                                  
#>  1 Caligus affinis   Pomatomus saltatrix Oldewage, W.H. & J.G. van As. (1989). …
#>  2 Caligus affinis   Umbrina cirrosa     Kensley, B. & J.R. Grindley. (1973). S…
#>  3 Caligus alalongae Thunnus alalunga    Raibaut, A., C. Combes & F. Benoit. (1…
#>  4 Caligus alalongae Thunnus obesus      Oldewage, W.H. & J.G. van As. (1989). …
#>  5 Caligus alalongae Thunnus obesus      Dippenaar, S.M. (2005). Reported sipho…
#>  6 Caligus alalongae Thunnus obesus      Capart, A. (1959). Copépodes parasites…
#>  7 Caligus belones   Belone belone       Cressey, R.F. & B.B. Collette. (1970).…
#>  8 Caligus belones   Belone belone       Holmes, J.M.C. (1998). A checklist of …
#>  9 Caligus belones   Belone svetovidovi  Holmes, J.M.C. (1998). A checklist of …
#> 10 Caligus belones   Belone belone       Dorman, J.A. (1991). Investigations in…
#> # ℹ 1,854 more rows