Find copepod occurrence data for given host fish. Given a fish family, genus or species returns a list of copepod occurrences on that host

Find copepod occurrence data for given host fish. Given a fish family, genus or species returns a list of copepod occurrences on that host

Usage

getFish(species = NULL, genus = NULL, family = NULL, citation = TRUE)

Arguments

species

Fish species to filter

genus

Fish genus to filter

family

Fish family to filter

citation

boolean. Should the output include the citation information? default is FALSE

Value

A data.frame with copepod fish interactions

Examples

getFish(genus = "Molva")
#> # A tibble: 14 × 3
#>    source_taxon_name       target_taxon_name study_citation                     
#>    <chr>                   <chr>             <chr>                              
#>  1 Cresseyus confusus      Molva molva       Holmes, J.M.C. & R.V. Gotto. (1992…
#>  2 Caligus curtus          Molva molva       Holmes, J.M.C. (1998). A checklist…
#>  3 Caligus curtus          Molva molva       Krøyer, H. (1843-1845). Danmarks F…
#>  4 Caligus elongatus       Molva molva       Boxshall, G.A. (1974). Infections …
#>  5 Lernaeocera branchialis Molva molva       Kabata, Z. (2003). Copepods Parasi…
#>  6 Lernaeocera lusci       Molva molva       Kabata, Z. (2003). Copepods Parasi…
#>  7 Sphyrion lumpi          Molva dypterygia  Priebe, K. (1987). Some fish speci…
#>  8 Sphyrion lumpi          Molva dypterygia  Ho, J.S. (1992). Does Sphyrion lum…
#>  9 Cresseyus confusus      Molva molva       Boxshall, G.A. (1974). Infections …
#> 10 Chondracanthus brotulae Molva molva       Ho, J.S. & J. Rokicki. (1987). Poe…
#> 11 Sarcotaces arcticus     Molva dypterygia  Kabata, Z. (1988). Copepoda and Br…
#> 12 Sarcotaces arcticus     Molva dypterygia  Hjort, J. (1895). Zur Anatomie und…
#> 13 Sarcotaces arcticus     Molva dypterygia  Collett, R. (1874). Meddelelse om …
#> 14 Cresseyus confusus      Molva molva       Ho, J.S. & C.L. Lin. (2006). A new…